feat: UIG-3270 - circulaire dependencies verbeterd
- GlobalStyles logica naar common verplaatst
- in lib/components geen '@domg-wc/components' imports, maar ze relatief importeren
- test utilities onder resources/utils-test om dependencies, die als neveneffect via barrel-file imports binnen komen, te vermijden
- swipeDetect probleem van de side-sheet opgelost
