+++ libs/common/jest.config.ts
@@ -12,6 +12,9 @@ const jestConfig: JestConfigWithTsJest = {
     },
     transformIgnorePatterns: [],
     testEnvironment: 'jsdom',
+    moduleNameMapper: {
+        "@resources/utils-test": '<rootDir>/../../resources/utils-test/index.ts',
+    }
 };
 
 export default jestConfig;
