+++ libs/common/src/styles/global-styles.cy.ts
