+++ libs/components/src/declare-module.d.ts
@@ -1,2 +1,3 @@
+declare module 'swipe-detect/dist/';
 declare module '*.mdx';
 declare module '*.css';
