+++ libs/components/src/proza-message/vl-proza-message.component.ts
@@ -1,5 +1,5 @@
 import { BaseElementOfType, registerWebComponents, webComponent } from '@domg-wc/common';
-import { VlButtonComponent } from '@domg-wc/components';
+import { VlButtonComponent } from '../button';
 import { VlTypography } from '../typography/vl-typography.component';
 import { VlProzaMessagePreloader } from './vl-proza-message-preloader.component';
 import elementStyles from './vl-proza-message.uig-css';
