+++ libs/components/src/rich-data-table/stories/vl-rich-data-table-pagination.stories-util.ts
@@ -1,4 +1,4 @@
-import { VlPagerComponent } from '@domg-wc/components';
+import { VlPagerComponent } from '../../pager';
 import richDataFilterPagerData from './vl-rich-data-table-pagination.stories-mock';
 
 export const paginationRichTableImplementation = () => {
