+++ libs/components/src/side-sheet/stories/vl-side-sheet.stories-utils.ts
@@ -1,4 +1,4 @@
-import { VlSideSheet } from '@domg-wc/components';
+import { VlSideSheet } from '../vl-side-sheet.component';
 
 export const sideSheetToggleImplementation = () => {
     let sideSheet: VlSideSheet;
