--- libs/components/src/side-sheet/swipe-detect.d.ts
@@ -1,4 +1,0 @@
-declare module 'swipe-detect/dist' {
-    const swipeDetect: any;
-    export default swipeDetect;
-}
