+++ libs/form/src/declare-module.d.ts
@@ -1,3 +1,4 @@
+declare module 'swipe-detect/dist/';
 declare module 'cleave.js';
 declare module 'dropzone';
 declare module '*.mdx';
