+++ libs/styles/src/base/body/vl-body.css.cy.ts
@@ -1,5 +1,5 @@
+import { GlobalStyles } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 
 describe('body styles', () => {
     beforeEach(() => {
