+++ libs/styles/src/base/link/vl-link.css.cy.ts
@@ -1,6 +1,5 @@
-import { hexToString } from '@domg-wc/common';
+import { GlobalStyles, hexToString } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 import { vlIconStyles } from '../icon/vl-icon.css';
 import { vlLinkStyles } from './vl-link.css';
