+++ libs/styles/src/base/mixin/vl-outlines.css.cy.ts
@@ -1,5 +1,5 @@
+import { GlobalStyles } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 import { vlFocusOutlineMixin } from './vl-outlines.css';
 
 describe('outline styles', () => {
