+++ libs/styles/src/layout/section/vl-section.css.cy.ts
@@ -1,5 +1,5 @@
+import { GlobalStyles } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 
 describe('section styles', () => {
     it('should style the defaults', () => {
