+++ libs/styles/src/layout/spacer/vl-spacer.css.cy.ts
@@ -1,5 +1,5 @@
+import { GlobalStyles } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 
 describe('spacer styles', () => {
     it('should have default spacer style', () => {
