+++ libs/styles/src/layout/stacked/vl-stacked.css.cy.ts
@@ -1,5 +1,5 @@
+import { GlobalStyles } from '@domg-wc/common';
 import { html } from 'lit';
-import { GlobalStyles } from '../../global-styles';
 
 describe('stacked styles', () => {
     const stackedLarge = html`
