+++ resources/utils-build/copy-common-js.mjs
@@ -2,7 +2,7 @@
 
 import { copyFiles } from './file-processors.mjs';
 
-copyFiles('build/tsc/lib-common', 'build/dist/libs/common', '');
+copyFiles('build/tsc/lib-common/common/src', 'build/dist/libs/common', '');
 copyFiles('libs/common/src', 'build/dist/libs/common', '.js');
 copyFiles('libs/common/src', 'build/dist/libs/common', '.css');
 copyFiles('libs/common', 'build/dist/libs/common', 'package.json');
