+++ resources/utils-test/index.ts
@@ -1,0 +1,1 @@
+export { buildSpan, buildDiv, buildLabel, buildData } from './html-element.builder';
